Use namespace for types
#19
