+++ src/getDate.ts
@@ -1,6 +1,6 @@
-import { Date, DateTime } from "./types";
+import { predate } from "./types";
 
-export function getDate(dateTime: DateTime): Date {
+export function getDate(dateTime: predate.DateTime): predate.Date {
   return {
     day: dateTime.day,
     month: dateTime.month,
