+++ src/getTime.ts
@@ -1,6 +1,6 @@
-import { DateTime, Time } from "./types";
+import { predate } from "./types";
 
-export function getTime(dateTime: DateTime): Time {
+export function getTime(dateTime: predate.DateTime): predate.Time {
   return {
     hours: dateTime.hours,
     minutes: dateTime.minutes,
