--- src/types/Date.ts
@@ -1,5 +1,0 @@
-export type Date = {
-  day: number;
-  month: number;
-  year: number;
-};
