--- src/types/DateTime.ts
@@ -1,3 +1,0 @@
-import { Date, Time } from ".";
-
-export type DateTime = Date & Time;
