--- src/types/Time.ts
@@ -1,6 +1,0 @@
-export type Time = {
-  hours: number;
-  minutes: number;
-  seconds?: number;
-  milliseconds?: number;
-};
